Specialize insertion logic for a and b children
In the IntSet implementation, we have a function insertNode to add new
nodes to the frontier. Each node that is deleted from the frontier can
have up to two child nodes that might be added to the frontier:
  * one node from advancing the as list
  * one node from advancing the bs list
insertNode is used to insert both types of nodes, and performs two
IntSet membership lookups for every inserted node.

This commit adds specialized insertChildA and insertChildB functions to
insert each kind of child node separately, and these functions should be
faster, as they perform only one IntSet membership test for every
inserted node.