Add mutable priority queue implementation
Add an ST-based mutable priority queue in Data.PQueue.Prio.Min.Mutable,
called MMinPQueue. Also add tests and benchmarks for MMinPQueue.

MMinPQueue is actually slower than the immutable priority queue from
pqueue, based on the current benchmarks. So, we don't implement an
applyMerge based on MMinPQueue. But we keep the implementation of
MMinPQueue around, in case we want to improve it later.