Only include necessary modules in library
There are multiple applyMerge implementations in src/, because we're
still trying to find the most performant way to implement applyMerge. We
don't want to include all these implementations in the apply-merge
library, because we only want to expose the current best implemetation
to users. However, we still want to test and benchmark all these
implementations.

So, we rework package.yaml so that the apply-merge library only exports
the current best applyMerge implementation, while the test and benchmark
suites still test and benchmark all the implementations.