Move implementations to ApplyMerge namespace
Previously, the different applyMerge implementations lived under
Data.List.ApplyMerge.*. But we plan to change the applyMerge
implementations to work on NonEmpty instead of List, and define the
canonical applyMerge List implementation in terms of the canonical
NonEmpty implementation.

Then if we were to keep the current namespacing, we would have to move
the applyMerge implementations to Data.List.ApplyMerge.NonEmpty.*, which
is a mouthful. Instead, we define a new namespace ApplyMerge.* for the
different implementations.