Configure CI to run on PRs and pushes to main
Previously, the CI was configured to run on pushes to all branches, and
on all pull requests. This caused pushes to a PR to cause all CI jobs to
run twice, once because they were a push, and once because the PR was
updated.

To fix this, we configure CI to run on all pull requests, and on pushes
only to main.