Use lts-21.25, corresponding to ghc-9.4.8
We were previously using the lts-22 series, which is for ghc-9.6.
ghc-9.6 is still receiving minor version updates, which means that we
have to bump the lts resolver often to keep using the latest minor
version. And it's important to use the latest minor version for
ergonomics, since HLS only supports the latest minor version of each ghc
major version, and commands like "ghcup install ghc 9.6" will install
the latest minor version.

Therefore, to avoid needing to bump the resolver too often, we set the
resolver to lts-21.25, which is the latest lts-21. This corresponds to
ghc-9.4.8, which is the latest minor version of ghc-9.4. Both lts-21 and
ghc-9.4 no longer receive updates, and ghc-9.4 is supported by HLS, and
ghc-9.4.8 is the recommended compiler version from ghcup. So now, we
don't have to bump the resolver too often, and we get good tooling
support.

In the future, when ghc-9.6 and lts-22 stop receiving updates, we can
bump the resolver to the latest lts-22.